AV1: Added a few changes to enable general testing (non-film-grain). (#383)
* * rocDecode/AV1: Added a few changes to enable general testing (non-film-grain).
 - Added show existing frame support.
 - Added load_previous() function.
 - Added more operations to reference frame update process.
 - Added reference frame loading process.
 - Added decode frame wrapup process.
 - Added DPB recycling.
 - Added DPB content logging funcion for debugging purposes.

* * rocDecode/AV1: Removed a line which was added accidentally in the previous commit.

+++ src/parser/av1_defines.h
@@ -298,7 +298,7 @@ typedef struct {
     uint32_t segmentation_temporal_update;
     uint32_t segmentation_update_data;
     uint32_t feature_enabled;
-    uint32_t feature_enabled_flags[MAX_SEGMENTS][SEG_LVL_MAX];
+    uint8_t  feature_enabled_flags[MAX_SEGMENTS][SEG_LVL_MAX];
     uint32_t feature_value;
     int16_t  feature_data[MAX_SEGMENTS][SEG_LVL_MAX];
     uint32_t seg_id_pre_skip;
@@ -322,9 +322,9 @@ typedef struct {
     uint32_t loop_filter_delta_enabled;
     uint32_t loop_filter_delta_update;
     uint32_t update_ref_delta;
-    uint32_t loop_filter_ref_deltas[TOTAL_REFS_PER_FRAME];
+    int32_t  loop_filter_ref_deltas[TOTAL_REFS_PER_FRAME];
     uint32_t update_mode_delta;
-    uint32_t loop_filter_mode_deltas[2];
+    int32_t  loop_filter_mode_deltas[2];
 } Av1LoopFilterParams;
 
 typedef struct {
